notion_extractor/requirements.txt
notion-client>=2.0.0
python-dotenv>=0.19.0
tqdm>=4.62.0
